requirements.txt
# Base on YOLOv5's requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Plotting ------------------------------------
pandas
seaborn

# Convertion Type
onnx
onnx-simplifier


# GUI + Topping
py-topping
PySimpleGUI
GitPython
google-cloud-storage
requests
sqlalchemy